+++ framework/grid/GridView.php
@@ -11,9 +11,7 @@ use Yii;
 use Closure;
 use yii\i18n\Formatter;
 use yii\base\InvalidConfigException;
-use yii\helpers\Url;
 use yii\helpers\Html;
-use yii\helpers\Json;
 use yii\widgets\BaseListView;
 use yii\base\Model;
 
@@ -279,20 +277,6 @@ class GridView extends BaseListView
     }
 
     /**
-     * Runs the widget.
-     */
-    public function run()
-    {
-        $id = $this->options['id'];
-        $options = Json::htmlEncode($this->getClientOptions());
-        $view = $this->getView();
-        GridViewAsset::register($view);
-        $view->registerJs("jQuery('#$id').yiiGridView($options);");
-
-        return parent::run();
-    }
-
-    /**
      * Renders validator errors of filter model.
      * @return string the rendering result.
      */
@@ -316,25 +300,6 @@ class GridView extends BaseListView
             default:
                 return parent::renderSection($name);
         }
-    }
-
-    /**
-     * Returns the options for the grid view JS widget.
-     * @return array the options
-     */
-    protected function getClientOptions()
-    {
-        $filterUrl = isset($this->filterUrl) ? $this->filterUrl : Yii::$app->request->url;
-        $id = $this->filterRowOptions['id'];
-        $filterSelector = "#$id input, #$id select";
-        if (isset($this->filterSelector)) {
-            $filterSelector .= ', ' . $this->filterSelector;
-        }
-
-        return [
-            'filterUrl' => Url::to($filterUrl),
-            'filterSelector' => $filterSelector,
-        ];
     }
 
     /**
