+++ framework/jquery/GridView.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\jquery;
+
+use Yii;
+use yii\helpers\Json;
+use yii\helpers\Url;
+
+/**
+ * GridView is an enhanced version of [[\yii\grid\GridView]], which allows automatic filter submission via jQuery component.
+ *
+ * @author Qiang Xue <qiang.xue@gmail.com>
+ * @author Paul Klimov <klimov.paul@gmail.com>
+ * @since 2.1
+ */
+class GridView extends \yii\grid\GridView
+{
+    /**
+     * Runs the widget.
+     */
+    public function run()
+    {
+        $id = $this->options['id'];
+        $options = Json::htmlEncode($this->getClientOptions());
+        $view = $this->getView();
+        GridViewAsset::register($view);
+        $view->registerJs("jQuery('#$id').yiiGridView($options);");
+
+        return parent::run();
+    }
+
+    /**
+     * Returns the options for the grid view JS widget.
+     * @return array the options
+     */
+    protected function getClientOptions()
+    {
+        $filterUrl = isset($this->filterUrl) ? $this->filterUrl : Yii::$app->request->url;
+        $id = $this->filterRowOptions['id'];
+        $filterSelector = "#$id input, #$id select";
+        if (isset($this->filterSelector)) {
+            $filterSelector .= ', ' . $this->filterSelector;
+        }
+
+        return [
+            'filterUrl' => Url::to($filterUrl),
+            'filterSelector' => $filterSelector,
+        ];
+    }
+}
