+++ framework/widgets/ActiveField.php
@@ -13,7 +13,6 @@ use yii\base\ErrorHandler;
 use yii\helpers\ArrayHelper;
 use yii\helpers\Html;
 use yii\base\Model;
-use yii\web\JsExpression;
 
 /**
  * ActiveField represents a form input field within an [[ActiveForm]].
@@ -96,49 +95,6 @@ class ActiveField extends Component
      * @see \yii\helpers\Html::renderTagAttributes() for details on how attributes are being rendered.
      */
     public $hintOptions = ['class' => 'hint-block'];
-    /**
-     * @var bool whether to enable client-side data validation.
-     * If not set, it will take the value of [[ActiveForm::enableClientValidation]].
-     */
-    public $enableClientValidation;
-    /**
-     * @var bool whether to enable AJAX-based data validation.
-     * If not set, it will take the value of [[ActiveForm::enableAjaxValidation]].
-     */
-    public $enableAjaxValidation;
-    /**
-     * @var bool whether to perform validation when the value of the input field is changed.
-     * If not set, it will take the value of [[ActiveForm::validateOnChange]].
-     */
-    public $validateOnChange;
-    /**
-     * @var bool whether to perform validation when the input field loses focus.
-     * If not set, it will take the value of [[ActiveForm::validateOnBlur]].
-     */
-    public $validateOnBlur;
-    /**
-     * @var bool whether to perform validation while the user is typing in the input field.
-     * If not set, it will take the value of [[ActiveForm::validateOnType]].
-     * @see validationDelay
-     */
-    public $validateOnType;
-    /**
-     * @var int number of milliseconds that the validation should be delayed when the user types in the field
-     * and [[validateOnType]] is set `true`.
-     * If not set, it will take the value of [[ActiveForm::validationDelay]].
-     */
-    public $validationDelay;
-    /**
-     * @var array the jQuery selectors for selecting the container, input and error tags.
-     * The array keys should be `container`, `input`, and/or `error`, and the array values
-     * are the corresponding selectors. For example, `['input' => '#my-input']`.
-     *
-     * The container selector is used under the context of the form, while the input and the error
-     * selectors are used under the context of the container.
-     *
-     * You normally do not need to set this property as the default selectors should work well for most cases.
-     */
-    public $selectors = [];
     /**
      * @var array different parts of the field (e.g. input, label). This will be used together with
      * [[template]] to generate the final field HTML code. The keys are the token names in [[template]],
@@ -226,13 +182,6 @@ class ActiveField extends Component
      */
     public function begin()
     {
-        if ($this->form->enableClientScript) {
-            $clientOptions = $this->getClientOptions();
-            if (!empty($clientOptions)) {
-                $this->form->attributes[] = $clientOptions;
-            }
-        }
-
         $inputID = $this->getInputId();
         $attribute = Html::getAttributeName($this->attribute);
         $options = $this->options;
@@ -741,102 +690,6 @@ class ActiveField extends Component
         if (!isset($this->labelOptions['for'])) {
             $this->labelOptions['for'] = $options['id'];
         }
-    }
-
-    /**
-     * Returns the JS options for the field.
-     * @return array the JS options.
-     */
-    protected function getClientOptions()
-    {
-        $attribute = Html::getAttributeName($this->attribute);
-        if (!in_array($attribute, $this->model->activeAttributes(), true)) {
-            return [];
-        }
-
-        $clientValidation = $this->isClientValidationEnabled();
-        $ajaxValidation = $this->isAjaxValidationEnabled();
-
-        if ($clientValidation) {
-            $validators = [];
-            foreach ($this->model->getActiveValidators($attribute) as $validator) {
-                /* @var $validator \yii\validators\Validator */
-                $js = $validator->clientValidateAttribute($this->model, $attribute, $this->form->getView());
-                if ($validator->enableClientValidation && $js != '') {
-                    if ($validator->whenClient !== null) {
-                        $js = "if (({$validator->whenClient})(attribute, value)) { $js }";
-                    }
-                    $validators[] = $js;
-                }
-            }
-        }
-
-        if (!$ajaxValidation && (!$clientValidation || empty($validators))) {
-            return [];
-        }
-
-        $options = [];
-
-        $inputID = $this->getInputId();
-        $options['id'] = Html::getInputId($this->model, $this->attribute);
-        $options['name'] = $this->attribute;
-
-        $options['container'] = isset($this->selectors['container']) ? $this->selectors['container'] : ".field-$inputID";
-        $options['input'] = isset($this->selectors['input']) ? $this->selectors['input'] : "#$inputID";
-        if (isset($this->selectors['error'])) {
-            $options['error'] = $this->selectors['error'];
-        } elseif (isset($this->errorOptions['class'])) {
-            $options['error'] = '.' . implode('.', preg_split('/\s+/', $this->errorOptions['class'], -1, PREG_SPLIT_NO_EMPTY));
-        } else {
-            $options['error'] = isset($this->errorOptions['tag']) ? $this->errorOptions['tag'] : 'span';
-        }
-
-        $options['encodeError'] = !isset($this->errorOptions['encode']) || $this->errorOptions['encode'];
-        if ($ajaxValidation) {
-            $options['enableAjaxValidation'] = true;
-        }
-        foreach (['validateOnChange', 'validateOnBlur', 'validateOnType', 'validationDelay'] as $name) {
-            $options[$name] = $this->$name === null ? $this->form->$name : $this->$name;
-        }
-
-        if (!empty($validators)) {
-            $options['validate'] = new JsExpression("function (attribute, value, messages, deferred, \$form) {" . implode('', $validators) . '}');
-        }
-
-        if ($this->addAriaAttributes === false) {
-            $options['updateAriaInvalid'] = false;
-        }
-
-        // only get the options that are different from the default ones (set in yii.activeForm.js)
-        return array_diff_assoc($options, [
-            'validateOnChange' => true,
-            'validateOnBlur' => true,
-            'validateOnType' => false,
-            'validationDelay' => 500,
-            'encodeError' => true,
-            'error' => '.help-block',
-            'updateAriaInvalid' => true,
-        ]);
-    }
-
-    /**
-     * Checks if client validation enabled for the field
-     * @return bool
-     * @since 2.0.11
-     */
-    protected function isClientValidationEnabled()
-    {
-        return $this->enableClientValidation || $this->enableClientValidation === null && $this->form->enableClientValidation;
-    }
-
-    /**
-     * Checks if ajax validation enabled for the field
-     * @return bool
-     * @since 2.0.11
-     */
-    protected function isAjaxValidationEnabled()
-    {
-        return $this->enableAjaxValidation || $this->enableAjaxValidation === null && $this->form->enableAjaxValidation;
     }
 
     /**
