+++ framework/widgets/ActiveForm.php
@@ -12,9 +12,7 @@ use yii\base\InvalidCallException;
 use yii\base\Widget;
 use yii\base\Model;
 use yii\helpers\ArrayHelper;
-use yii\helpers\Url;
 use yii\helpers\Html;
-use yii\helpers\Json;
 
 /**
  * ActiveForm is a widget that builds an interactive HTML form for one or multiple data models.
@@ -92,83 +90,6 @@ class ActiveForm extends Widget
      * @var string the CSS class that is added to a field container when the associated attribute is successfully validated.
      */
     public $successCssClass = 'has-success';
-    /**
-     * @var string the CSS class that is added to a field container when the associated attribute is being validated.
-     */
-    public $validatingCssClass = 'validating';
-    /**
-     * @var bool whether to enable client-side data validation.
-     * If [[ActiveField::enableClientValidation]] is set, its value will take precedence for that input field.
-     */
-    public $enableClientValidation = true;
-    /**
-     * @var bool whether to enable AJAX-based data validation.
-     * If [[ActiveField::enableAjaxValidation]] is set, its value will take precedence for that input field.
-     */
-    public $enableAjaxValidation = false;
-    /**
-     * @var bool whether to hook up `yii.activeForm` JavaScript plugin.
-     * This property must be set `true` if you want to support client validation and/or AJAX validation, or if you
-     * want to take advantage of the `yii.activeForm` plugin. When this is `false`, the form will not generate
-     * any JavaScript.
-     */
-    public $enableClientScript = true;
-    /**
-     * @var array|string the URL for performing AJAX-based validation. This property will be processed by
-     * [[Url::to()]]. Please refer to [[Url::to()]] for more details on how to configure this property.
-     * If this property is not set, it will take the value of the form's action attribute.
-     */
-    public $validationUrl;
-    /**
-     * @var bool whether to perform validation when the form is submitted.
-     */
-    public $validateOnSubmit = true;
-    /**
-     * @var bool whether to perform validation when the value of an input field is changed.
-     * If [[ActiveField::validateOnChange]] is set, its value will take precedence for that input field.
-     */
-    public $validateOnChange = true;
-    /**
-     * @var bool whether to perform validation when an input field loses focus.
-     * If [[ActiveField::$validateOnBlur]] is set, its value will take precedence for that input field.
-     */
-    public $validateOnBlur = true;
-    /**
-     * @var bool whether to perform validation while the user is typing in an input field.
-     * If [[ActiveField::validateOnType]] is set, its value will take precedence for that input field.
-     * @see validationDelay
-     */
-    public $validateOnType = false;
-    /**
-     * @var int number of milliseconds that the validation should be delayed when the user types in the field
-     * and [[validateOnType]] is set `true`.
-     * If [[ActiveField::validationDelay]] is set, its value will take precedence for that input field.
-     */
-    public $validationDelay = 500;
-    /**
-     * @var string the name of the GET parameter indicating the validation request is an AJAX request.
-     */
-    public $ajaxParam = 'ajax';
-    /**
-     * @var string the type of data that you're expecting back from the server.
-     */
-    public $ajaxDataType = 'json';
-    /**
-     * @var bool whether to scroll to the first error after validation.
-     * @since 2.0.6
-     */
-    public $scrollToError = true;
-    /**
-     * @var int offset in pixels that should be added when scrolling to the first error.
-     * @since 2.0.11
-     */
-    public $scrollToErrorOffset = 0;
-    /**
-     * @var array the client validation options for individual attributes. Each element of the array
-     * represents the validation options for a particular attribute.
-     * @internal
-     */
-    public $attributes = [];
 
     /**
      * @var ActiveField[] the ActiveField objects that are currently active
